+++ src/danmaku_sender/ui/common/notification.py
@@ -2,9 +2,9 @@ import logging
 import platform
 import threading
 
-from .path_utils import get_assets_path
+from ..framework.path_utils import get_assets_path
 
-from ..config.app_meta import AppInfo
+from ...config.app_meta import AppInfo
 
 
 logger = logging.getLogger("App.System.Notify")
