+++ src/danmaku_sender/ui/controllers/sender_controller.py
@@ -6,6 +6,8 @@ from PySide6.QtCore import QObject, Signal, Slot
 
 from danmaku_sender.ui.framework.concurrency import WorkerThread, PoolTask
 
+from .system_utils import KeepSystemAwake
+
 from danmaku_sender.repo.history_manager import HistoryManager
 from danmaku_sender.service.sender import DanmakuScheduler, DanmakuExecutor, SendingContext, SendJob, DelayManager
 from danmaku_sender.types.models.danmaku import Danmaku
@@ -16,7 +18,6 @@ from danmaku_sender.runtime.app_state import AppState
 from danmaku_sender.service.danmaku_parser import DanmakuParser
 from danmaku_sender.service.danmaku_exporter import create_xml_from_danmakus
 from danmaku_sender.repo.bili_api_client import BiliApiClient
-from danmaku_sender.utils.system_utils import KeepSystemAwake
 
 
 logger = logging.getLogger("App.System.SenderController")
