+++ src/danmaku_sender/ui/framework/style_loader.py
@@ -7,7 +7,7 @@ from PySide6.QtWidgets import QApplication
 from .image_processor import QtImageProcessor
 from danmaku_sender.runtime.theme_manager import ThemeManager
 
-from ...utils.path_utils import get_assets_path
+from .path_utils import get_assets_path
 
 
 logger = logging.getLogger("App.UI.StyleLoader")
