+++ src/danmaku_sender/ui/sender_page.py
@@ -21,7 +21,7 @@ from danmaku_sender.types.models.common import VideoTarget, UnsentDanmakusRecord
 from danmaku_sender.runtime.app_state import AppState
 from danmaku_sender.utils.string_utils import parse_bilibili_link
 from danmaku_sender.utils.time_utils import format_duration
-from danmaku_sender.utils.notification_utils import send_windows_notification
+from danmaku_sender.ui.common.notification import send_windows_notification
 
 
 class SenderPage(QWidget):
